Modifying e2e yaml and adding test logs upload

+++ e2e/logs/upload_logs.sh
@@ -25,7 +25,7 @@ aws configure set default.output json
 LOCAL_LOGS_DIR="$RUN_ID"
 
 for file in *.log; do
-    if [ -f "$file" ]; then
+    if [ -f "$file" ] && [ -s "$file" ]; then
         filename=$(basename "$file")
         aws s3 cp "$file" "s3://$S3_BUCKET/$LOCAL_LOGS_DIR/$filename"
     fi
